chore: moved icons and buttons
