+++ src/components/atoms/buttons/follow-button-user/type.ts
