--- src/components/atoms/task-icon/index.ts
@@ -1,1 +1,0 @@
-export * from './task-icon';
