+++ src/components/features/explore/tabs/people-tab/user-cell.tsx
@@ -12,7 +12,7 @@ import { Box } from '@mui/system';
 
 const FollowButtonUser = dynamic<any>(
   () =>
-    import('@/components/atoms/follow-button-user').then(
+    import('@/components/atoms/buttons/follow-button-user').then(
       (mod) => mod.FollowButtonUser
     ),
   {
