+++ src/components/features/gates/create/tasks/file-link-task/file-link-task.tsx
@@ -1,7 +1,7 @@
 import { useEffect, useState } from 'react';
 
 import { CircleWithNumber } from '@/components/atoms/circle-with-number';
-import { TaskIcon } from '@/components/atoms/task-icon';
+import { TaskIcon } from '@/components/atoms/icons/task-icon';
 import {
   CreateGateData,
   FileTaskDataError,
