+++ src/components/features/gates/create/tasks/hold-token-task/hold-token-task.tsx
@@ -1,6 +1,6 @@
 import { useEffect, useState } from 'react';
 
-import { TaskIcon } from '@/components/atoms/task-icon';
+import { TaskIcon } from '@/components/atoms/icons/task-icon';
 import {
   CreateGateData,
   HoldTokenDataError,
