+++ src/components/features/gates/create/tasks/snapshot-task/snapshot-task.tsx
@@ -1,6 +1,6 @@
 import { useEffect, useState } from 'react';
 
-import { TaskIcon } from '@/components/atoms/task-icon';
+import { TaskIcon } from '@/components/atoms/icons/task-icon';
 import {
   CreateGateData,
   SnapshotDataError,
