+++ src/components/features/gates/create/tasks/twitter-retweet/twitter-retweet.tsx
@@ -1,6 +1,6 @@
 import { useCallback, useEffect, useState } from 'react';
 
-import { TaskIcon } from '@/components/atoms/task-icon';
+import { TaskIcon } from '@/components/atoms/icons/task-icon';
 import {
   CreateGateData,
   TwitterRetweetDataError,
