+++ src/components/features/gates/create/tasks/twitter-tweet/twitter-tweet.tsx
@@ -1,6 +1,6 @@
 import { useEffect, useRef, useState } from 'react';
 
-import { TaskIcon } from '@/components/atoms/task-icon';
+import { TaskIcon } from '@/components/atoms/icons/task-icon';
 import {
   CreateGateData,
   TwitterTweetDataError,
