+++ src/components/features/search/tabs/people-tab/table-view.tsx
@@ -17,7 +17,7 @@ import TableRow from '@mui/material/TableRow';
 
 const FollowButtonUser = dynamic<any>(
   () =>
-    import('@/components/atoms/follow-button-user').then(
+    import('@/components/atoms/buttons/follow-button-user').then(
       (mod) => mod.FollowButtonUser
     ),
   {
